requirements.txt
Django==1.9
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
Markdown==2.6.8
Pillow==3.1.0
PyJWT==1.5.2
